+++ swoole_http.h
@@ -226,7 +226,7 @@ static int http_parse_set_cookies(const char *at, size_t length, zval *cookies, 
     }
     if (key_len > SW_HTTP_COOKIE_KEYLEN)
     {
-        swWarn("cookie[%.8s...] name length %d is exceed the max name len %d", key, key_len, SW_HTTP_COOKIE_KEYLEN);
+        swWarn("cookie[%.8s...] name length %d is exceed the max name len %d", key, (int) key_len, SW_HTTP_COOKIE_KEYLEN);
         return SW_ERR;
     }
     add_assoc_stringl_ex(set_cookie_headers, key, key_len, (char *) at, length);
